Fix subverter chips not having an announcement
