+++ Content.Client/_ES/Stagehand/ESStagehandNotificationsSystem.cs
@@ -1,0 +1,5 @@
+using Content.Shared._ES.Stagehand;
+
+namespace Content.Client._ES.Stagehand;
+
+public sealed class ESStagehandNotificationsSystem : ESSharedStagehandNotificationsSystem;
