+++ Content.Server/_ES/SecretIdentity/Tragedian/ESCallStagehandsActionSystem.cs
@@ -1,5 +1,6 @@
 using Content.Server._ES.Stagehand;
 using Content.Shared._ES.SecretIdentity.Tragedian;
+using Content.Shared._ES.Stagehand;
 using Content.Shared._ES.Stagehand.Components;
 using Content.Shared.Follower;
