+++ Resources/Locale/en-US/_ES/secret-identities/traitor/subverter.ftl
@@ -1,5 +1,6 @@
 ﻿es-subverter-chip-used = It's already been used!
-es-subverter-chip-implanting = Implanting subversion chip!
+es-subverter-chip-implanting = Implanting subversion chip!\
+es-subverter-chip-announcement = {$player} used {INDEFINITE($item)} {$item} on {$target}.
 es-subverter-chip-examined-usable = [color=green]This chip has not been used.[/color]
 es-subverter-chip-examined-used = [color=red]This chip has been used![/color]
 es-subverter-chip-not-crit = Target aware!
